# src/main/resources/public/api.yml
openapi: 3.0.0
info:
  title: Bruse API
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 0.0.1
servers:
  - url: https://api.bruse.dev/
    description: Main (production) server
paths:
  /task/request:
    post:
      summary: Requests a task
      operationId: requestTask
      tags:
        - Bruse
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequestDto'
      responses:
        '200':
          description: A new task was found. This is returned to requester.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '400':
          description: The provided request body was not accepted.
        '404':
          description: No new task was found.
        '500':
          description: Something went wrong when trying to retrieve a task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /task/content/{contentId}:
    get:
      summary: Get task content
      operationId: getTaskContent
      parameters:
        - name: contentId
          in: path
          schema:
            type: string
          required: true
      tags:
        - Bruse
      responses:
        '200':
          description: Returns the requested task content.
          content:
            image/*:
              schema:
                type: string
                format: binary
            text/html:
              schema:
                type: string
        '308':
          description: Returns redirect information to the given task content.
        '400':
          description: The provided content id was not accepted.
components:
  schemas:
    TaskRequestDto:
      type: object
      properties:
        gamepin:
          type: string
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
      required:
        - gamepin
        - lat
        - lon
    TaskDto:
      type: object
      properties:
        task_id:
          type: string
        latlon_coordinates:
          type: array
          maxLength: 2
          minLength: 2
          items:
            type: number
            format: double
        acceptance_radius:
          type: integer
        task_content_id:
          type: string
    ErrorResponseDto:
      type: object
      properties:
        error_message:
          type: string